+++ miles/backends/megatron_utils/update_weight/update_weight_from_tensor.py
@@ -14,11 +14,11 @@ from miles.backends.megatron_utils.lora_utils import LORA_ADAPTER_NAME, build_lo
 from miles.utils.distributed_utils import get_gloo_group
 
 from ..sglang import FlattenedTensorBucket, MultiprocessingSerializer
+from .common import post_process_weights
 from .hf_weight_iterator_base import HfWeightIteratorBase
-from .update_weight_from_distributed import (
+from .update_weight_from_distributed.broadcast import (
     connect_rollout_engines_from_distributed,
     disconnect_rollout_engines_from_distributed,
-    post_process_weights,
     update_weights_from_distributed,
 )
